+++ NIPTool/load/batch.py
@@ -1,8 +1,9 @@
 import logging
 from NIPTool.build.document import build_sample, build_batch
 from NIPTool.models.validation import requiered_fields
-from pathlib import Path
-from typing import Optional
+from NIPTool.parse.batch import pars_segmental_calls
+from NIPTool.exeptions import InvalidFileError
+
 
 LOG = logging.getLogger(__name__)
 
@@ -16,28 +17,12 @@ def check_requiered_fields(document):
     return True
 
 
-def pars_segmental_calls(segmental_calls_path: Optional[str]) -> dict:
-    """Builds a dict with segmental calls bed files.
-        key: sample ids
-        value: bed file path"""
-
-    segmental_calls_dir = Path(segmental_calls_path)
-    segmental_calls = {}
-    if not segmental_calls_dir.exists():
-        LOG.info('Segmental Calls file path missing.')
-        return segmental_calls
-
-    for file in segmental_calls_dir.iterdir():
-        if file.suffix == '.bed':
-            sample_id = file.name.split('.')[0]
-            segmental_calls[sample_id] = str(file.absolute())
-
-    return segmental_calls
-
 
 def load_batch(adapter, batch_data: dict, request_data: dict) -> None:
     """Function to load data from fluffy result file."""
 
+    if not check_requiered_fields(batch_data):
+        raise InvalidFileError(message="Could not load data from result file. Required fields missing.")
     mongo_batch = build_batch(batch_data, request_data)
     adapter.add_or_update_document(mongo_batch, adapter.batch_collection)
 
@@ -48,7 +33,7 @@ def load_samples(adapter, batch_data: list, request_data: dict) -> None:
     segmental_calls = pars_segmental_calls(segmental_calls_path=request_data.get('segmental_calls'))
     for sample in batch_data:
         if not check_requiered_fields(sample):
-            continue
+            raise InvalidFileError(message="Could not load data from result file. Required fields missing.")
         sample_id = sample["SampleID"]
         segmental_calls_path = segmental_calls.get(sample_id)
         mongo_sample = build_sample(sample_data=sample, segmental_calls=segmental_calls_path, sample_id=sample_id)
