--- tests/fixtures/not_valid_concentrations.yaml
@@ -1,1 +1,0 @@
-{ "2020-07547-02": 12, "2020-07649-02": 12, "2020-07455-02": 12
