--- tests/fixtures/valid_concentrations.yaml
@@ -1,1 +1,0 @@
-{ "s1": 12, "s2": 12, "s3": 12 }
